+++ core-bundle/src/Security/Authentication/Provider/AuthenticationProvider.php
@@ -12,13 +12,9 @@ declare(strict_types=1);
 
 namespace Contao\CoreBundle\Security\Authentication\Provider;
 
-use Contao\Config;
 use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
 use Contao\CoreBundle\Security\Exception\LockedException;
-use Contao\FrontendUser;
-use Contao\Idna;
 use Contao\User;
-use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Security\Core\Authentication\Provider\DaoAuthenticationProvider;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
@@ -27,7 +23,6 @@ use Symfony\Component\Security\Core\Exception\BadCredentialsException;
 use Symfony\Component\Security\Core\User\UserCheckerInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Translation\TranslatorInterface;
 
 class AuthenticationProvider extends DaoAuthenticationProvider
 {
@@ -35,21 +30,6 @@ class AuthenticationProvider extends DaoAuthenticationProvider
      * @var ContaoFrameworkInterface
      */
     private $framework;
-
-    /**
-     * @var TranslatorInterface
-     */
-    private $translator;
-
-    /**
-     * @var RequestStack
-     */
-    private $requestStack;
-
-    /**
-     * @var \Swift_Mailer
-     */
-    private $mailer;
 
     /**
      * @var array
@@ -62,19 +42,13 @@ class AuthenticationProvider extends DaoAuthenticationProvider
      * @param string                   $providerKey
      * @param EncoderFactoryInterface  $encoderFactory
      * @param ContaoFrameworkInterface $framework
-     * @param TranslatorInterface      $translator
-     * @param RequestStack             $requestStack
-     * @param \Swift_Mailer            $mailer
      * @param array                    $options
      */
-    public function __construct(UserProviderInterface $userProvider, UserCheckerInterface $userChecker, $providerKey, EncoderFactoryInterface $encoderFactory, ContaoFrameworkInterface $framework, TranslatorInterface $translator, RequestStack $requestStack, \Swift_Mailer $mailer, array $options = [])
+    public function __construct(UserProviderInterface $userProvider, UserCheckerInterface $userChecker, $providerKey, EncoderFactoryInterface $encoderFactory, ContaoFrameworkInterface $framework, array $options = [])
     {
         parent::__construct($userProvider, $userChecker, $providerKey, $encoderFactory, false);
 
         $this->framework = $framework;
-        $this->translator = $translator;
-        $this->requestStack = $requestStack;
-        $this->mailer = $mailer;
         $this->options = array_merge(['login_attempts' => 3, 'lock_period' => 300], $options);
     }
 
@@ -134,8 +108,6 @@ class AuthenticationProvider extends DaoAuthenticationProvider
         $lockedSeconds = $user->locked - time();
         $lockedMinutes = (int) ceil($lockedSeconds / 60);
 
-        $this->sendLockedEmail($user, $lockedMinutes);
-
         $exception = new LockedException(
             $lockedSeconds,
             sprintf('User "%s" has been locked for %s minutes', $user->username, $lockedMinutes),
@@ -146,56 +118,6 @@ class AuthenticationProvider extends DaoAuthenticationProvider
         $exception->setUser($user);
 
         return $exception;
-    }
-
-    /**
-     * Notifies the administrator of the locked account.
-     *
-     * @param User $user
-     * @param int  $lockedMinutes
-     *
-     * @throws \RuntimeException
-     */
-    private function sendLockedEmail(User $user, int $lockedMinutes): void
-    {
-        $this->framework->initialize();
-
-        /** @var Config $config */
-        $config = $this->framework->getAdapter(Config::class);
-
-        if ($adminEmail = $config->get('adminEmail')) {
-            $request = $this->requestStack->getMasterRequest();
-
-            if (null === $request) {
-                throw new \RuntimeException('The request stack did not contain a request');
-            }
-
-            $realName = $user->name;
-
-            if ($user instanceof FrontendUser) {
-                $realName = sprintf('%s %s', $user->firstname, $user->lastname);
-            }
-
-            $website = Idna::decode($request->getSchemeAndHttpHost());
-            $subject = $this->translator->trans('MSC.lockedAccount.0', [], 'contao_default');
-
-            $body = $this->translator->trans(
-                'MSC.lockedAccount.1',
-                [$user->username, $realName, $website, $lockedMinutes],
-                'contao_default'
-            );
-
-            $email = new \Swift_Message();
-
-            $email
-                ->setFrom($adminEmail)
-                ->setTo($adminEmail)
-                ->setSubject($subject)
-                ->setBody($body, 'text/plain')
-            ;
-
-            $this->mailer->send($email);
-        }
     }
 
     /**