+++ core-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## DEV
 
+ * Do not notify the admin via e-mail if an account is locked (see #7728).
  * Improve cache busting of the .css and .js files (see #1404).